confirmation on deleting

+++ client/src/components/Car.js
@@ -1,5 +1,7 @@
 import React, { useState, useEffect } from 'react';
-import { Container, Table, Button, Row, Col, Modal, Form, Spinner } from 'react-bootstrap'
+import { Container, Table, Button, Row, Col, Modal, Form, Spinner } from 'react-bootstrap';
+import Swal from 'sweetalert2';
+
 import RowTable from './RowTableModelCar';
 import axios from '../api/database';
 import Toast from './ToastComponent';
@@ -74,21 +76,32 @@ function Car (props) {
     }
 
     function deleteData(id){
-        axios.delete(`/cars/${id}`, { headers: { token:localStorage.getItem('token')}})
-            .then(({data}) => {
-                let tempTable = rowTable.filter((el,index) => {
-                    return el._id !== id
-                })
-                setRowTable(tempTable)
-                setTextToast('delete success')
-                setStatusToast(true)
-                setShowToast(true)
-            })
-            .catch(err => {
-                setTextToast(err.response.data.message)
-                setStatusToast(false)
-                setShowToast(true)
-            })
+        Swal.fire({
+            title: 'Delete data?',
+            text: "You are about to delete selected data, are you sure?",
+            icon: 'question',
+            showCancelButton: true,
+            confirmButtonText: 'Yes',
+            cancelButtonText: 'No'
+        }).then((result) => {
+            if (result.value) {
+                axios.delete(`/cars/${id}`, { headers: { token:localStorage.getItem('token')}})
+                    .then(({data}) => {
+                        let tempTable = rowTable.filter((el,index) => {
+                            return el._id !== id
+                        })
+                        setRowTable(tempTable)
+                        setTextToast('delete success')
+                        setStatusToast(true)
+                        setShowToast(true)
+                    })
+                    .catch(err => {
+                        setTextToast(err.response.data.message)
+                        setStatusToast(false)
+                        setShowToast(true)
+                    })
+            }
+        });
     }
 
     function handleClose() {
